+++ CHANGES.rst
@@ -5,7 +5,10 @@ History
 1.1 (unreleased)
 ----------------
 
-- Cleanup tests
+- Add maintenance utilities.
+  [rnix, 2014-05-13]
+
+- Cleanup tests.
   [rnix, 2014-05-13]
 
 1.0